+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+## [v3.4.0-beta.7]
+  - 📅 2025-03-28
   - 💄 [@mixte/snippets] 使 `Ant Design X` 的 `Bubble` 支持通过 `content` 传递 Vue 组件实现自定义渲染内容 *
   - 💄 [@mixte/snippets] 使 `Ant Design X` 的 `BubbleList` 支持通过 `items[].content` 传递 Vue 组件实现自定义渲染内容 *
   - 🐞 [@mixte/snippets] 修复 `Ant Design X` 的 `Conversations` 组件监听 `active-change` 事件报错的问题 *
@@ -119,7 +122,8 @@ CHANGELOG 图标规范
 <hr>
 <br>
 
-[Unreleased]: https://github.com/MoomFE/mixte/compare/v3.4.0-beta.6...HEAD
+[Unreleased]: https://github.com/MoomFE/mixte/compare/v3.4.0-beta.7...HEAD
+[v3.4.0-beta.7]: https://github.com/MoomFE/mixte/releases/tag/v3.4.0-beta.7
 [v3.4.0-beta.6]: https://github.com/MoomFE/mixte/releases/tag/v3.4.0-beta.6
 [v3.4.0-beta.5]: https://github.com/MoomFE/mixte/releases/tag/v3.4.0-beta.5
 [v3.4.0-beta.4]: https://github.com/MoomFE/mixte/releases/tag/v3.4.0-beta.4